% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{actg019}
\alias{actg019}
\title{AIDS Clinical Trial ACTG019 (1990).}
\format{
A data frame with 404 rows and 4 variables:
\describe{
\item{outcome}{binary variable with 1 indicating death, development of AIDS or ARC and 0 otherwise}
\item{age}{patient age in years}
\item{race}{binary variable with 1 indicating white and 0 otherwise}
\item{T4count}{CD4 cell count (cell count per cubicmillimetre of serum)}
}
}
\source{
Chen, Ming-Hui, et al. "Prior Elicitation, Variable Selection and Bayesian Computation for Logistic Regression Models." Journal of the Royal Statistical Society. Series B, vol. 61, no. 1, 1999, pp. 223-242.
}
\usage{
actg019
}
\description{
A dataset containing the ACTG019 clinical trial placebo group data (1990) in adults with asymptomatic HIV.
}
\keyword{datasets}
